fix: correct minor errors in the codebase
