+++ stories/Page.tsx
@@ -1,23 +1,12 @@
 import React from 'react';
 
-import { Header } from './Header';
+import Header from '../components/header';
 import './page.css';
 
-type User = {
-  name: string;
-};
-
 export const Page: React.FC = () => {
-  const [user, setUser] = React.useState<User>();
-
   return (
     <article>
-      <Header
-        user={user}
-        onLogin={() => setUser({ name: 'Jane Doe' })}
-        onLogout={() => setUser(undefined)}
-        onCreateAccount={() => setUser({ name: 'Jane Doe' })}
-      />
+      <Header/>
 
       <section className="storybook-page">
         <h2>Pages in Storybook</h2>
